==> jupyter/02_summary.ipynb <==
### Creating summaries of given csv file

import pandas as pd
import transformers
import torch

undisputed_best_model = transformers.MBartForConditionalGeneration.from_pretrained(
    "ml6team/mbart-large-cc25-cnn-dailymail-nl-finetune"
)
tokenizer = transformers.MBartTokenizer.from_pretrained("facebook/mbart-large-cc25")
summarization_pipeline = transformers.pipeline(
    task="summarization",
    model=undisputed_best_model,
    tokenizer=tokenizer,
)
summarization_pipeline.model.config.decoder_start_token_id = tokenizer.lang_code_to_id[
    "nl_XX"
]


torch.cuda.is_available()


torch.cuda.device_count()


undisputed_best_model.to('cuda')

df=pd.read_csv("../data_01.csv",sep="|")

articles=df['nl_body']
articles=articles[0:1]

summary=[]
#Creating the summary with pretrained model 
counter=0
t=None
for text in articles:
    

    t= summarization_pipeline(
        text,
        do_sample=True,
        top_p=0.75,
        top_k=50,
        num_beams=4,
        min_length=50,
        early_stopping=True,
        truncation=True,
    )[0]["summary_text"]


    counter=counter+1
    print(counter)

df['nl_sum']=summary

df.to_csv("data_02.csv",encoding="utf-8-sig",sep="|")
