==> jupyter/03_keywords.ipynb <==
### Finding keywords

import spacy
import pandas as pd
KEYWORDS = 'belasting bedrijfsvoorheffing aanslagjaar btw inkom'

nlp = spacy.load("nl_core_news_lg")
data=pd.read_csv('data_02.csv',sep="|")

data.head(3)

summ=data['nl_body']

main_keywords=nlp(KEYWORDS)
keys=[]
for _a in main_keywords:
    for article in summ:
        text=article.lower()
        docarticle=nlp(text)
        for token in docarticle:
            if _a.similarity(token)>0.75:
                keys.append(token)
          
dfkeys=pd.DataFrame(keys)
dfkeys.to_csv("keywords.csv")       

#AFTER SAVING THE KEYWORDS WE MUST DECIDE WHICH ONES TO USE AND WHICH ONES TO DELETE

from collections import Counter
items=[]
for a in keys:
    items.append(a.lower_)

freq_word=Counter(items)
freq_word.most_common(50)

itemsset=set(items)
itemsset
