==> jupyter/05_forsql.ipynb <==
### Writing data to sql

import sqlite3
import pandas as pd


con = sqlite3.connect("data.sqlite")

df=pd.read_csv("../data_03.csv",sep="|")

df.head(3)

df.to_sql("data",con=con,if_exists="replace",index=False)
con.close()

df = pd.read_sql_query("SELECT * from data", con)
con.close()

df.columns
    

df.rename(columns = {'Date':'date', 'Numac':'numac','Link NL':'link','nl_body':'nl_text','nltags':'nl_tags'}, inplace = True)


df.drop(['Title','Link FR','fr_body'],axis=1,inplace=True)

con2=sqlite3.connect("./btax/db.sqlite3")


df.to_sql('taxtag_article',con=con2,if_exists="replace")

con2.close()

df.to_csv('data_tagged.csv',index=False,encoding='utf-8-sig')

z=df["nl_body"].to_list()
z[0]

df[['NL ','nltags']].head(20)
q=df['nltags'].to_list()
q

#Delete German text
to_delete='De hierna volgende tekst is de Duitse vertaling'
#df.drop(df[df['nl_body'] < 50].index, inplace=True)

df=df[df["nl_body"].str.contains(to_delete)==False]
