==> jupyter/06_app.ipynb <==
### Testing different aproaches to use in django app

import numpy as np
import pandas as pd
import re

import requests
from bs4 import BeautifulSoup


import string
import unidecode


import transformers
import spacy


import torch

DATE='2022-06-09'

MAINLINK='https://www.ejustice.just.fgov.be/cgi/summary_body.pl?language=nl&pub_date='
DETAILLINK='https://www.ejustice.just.fgov.be/cgi/article_body.pl?language=nl&caller=summary&pub_date='



res = requests.get(MAINLINK + DATE)



def get_numac_numbers(res:str):
    _numacs=[]

    soup = BeautifulSoup(res, 'html.parser')

    try:
        value = soup.find_all('input', {'name': 'numac'})
        
    except Exception as e:
        print("Got unhandled exception %s" % str(e))

    for v in value:
        _numacs.append(v['value'].strip())
    return _numacs



def create_numac_links(_numacs:list):
    links=[]

    for _a in _numacs:
        link = f"{DETAILLINK}{DATE}&numac={_a}"
        links.append(link)
    return links

numacs=get_numac_numbers(res.text)
numac_links=create_numac_links(numacs)
numac_links

#Cleaning the scraped text
def clean(_a:str):
    d=re.sub(r'(?<=[.,;,:])(?=[^\s])', r' ', _a)
    
    document_test= unidecode.unidecode(d)
    document_test = document_test.replace('\\n', ' ').replace('\n', ' ').replace('\t',' ').replace('\\', ' ').replace('. com', '.com')

    pattern = re.compile(r'\s+') 
    Without_whitespace = re.sub(pattern, ' ', document_test)
    # There are some instances where there is no space after '?' & ')', 
    # I am replacing these with one space so that It will not consider two words as one token.
    document_test = Without_whitespace.replace('?', ' ? ').replace(')', ') ')
    
 
    document_test = re.sub(r"[^a-zA-Z0-9:$-,%.?!]+", ' ', document_test) 
   
    # Remove Mentions
    document_test = re.sub(r'@\w+', '', document_test)

    #Not necessary; 
    #document_test = re.sub(r"[^a-zA-Z:$-,%.?!]+", ' ', document_test)

    return document_test

#Scrape the article 

def scrape_numac(_numac_links:list):
    _count=0 # to check which line
    nl_list=[]
    for a in _numac_links:
        _count+=1
        res = requests.get(a)
        soup = BeautifulSoup(res.text, 'html.parser')
        for sup in soup.find_all('sup'):
            sup.unwrap()
    
        
        text=soup.text
        text=text.replace('\n',"")
        lst=text.split('Numac :')[1].split(text.split('Numac :')[2])
    

        article=lst[1].split('begin eerste woord laatste')[0].strip()
        article=clean(article)
        nl_list.append(article)
        print(_count)
    return nl_list
    

nl_list=scrape_numac(numac_links)

#SUMMARY
from transformers import MBartConfig


undisputed_best_model = transformers.MBartForConditionalGeneration.from_pretrained(
    "ml6team/mbart-large-cc25-cnn-dailymail-nl-finetune")




tokenizer = transformers.MBartTokenizer.from_pretrained("facebook/mbart-large-cc25")
summarization_pipeline = transformers.pipeline(
    task="summarization",
    model=undisputed_best_model,
    tokenizer=tokenizer,
)
summarization_pipeline.model.config.decoder_start_token_id = tokenizer.lang_code_to_id[
    "nl_XX"
]


#Creating the summary with pretrained model and saving to a csv file in case an error and appending to dataframe 

def summarise(_nl_list:list):
    counter=0
    summary=[]
    for text in _nl_list:
        

        t= summarization_pipeline(
            text,
            do_sample=True,
            top_p=0.75,
            top_k=50,
            num_beams=4,
            min_length=50,
            early_stopping=True,
            truncation=True,
        )[0]["summary_text"]

        summary.append(t)

        counter=counter+1
        print(counter)
    return summary

summary=summarise(nl_list)

### Tagging

summary

SIMILARITY=0.80
tags='aanslagjaar arbeidsongeschiktheidsuitkeringen bedrijfsinkomsten bedrijfskosten bedrijfstoeslag bedrijfsvoorheffing belasting belastingverdragen belastingverhoging'\
     'belastingvermindering belastingvoet belastingvoordeel belastingvrije beroepsinkomsten beroepskosten bezoldiging btw derdebetalersregeling dienstverplichtingen erfbelasting'\
     'financieringskosten heffing inkomsten inkomstenderving investeringsaftrek kapitaalaflossingen kapitaalvermindering kostenvermindering omzetbelasting personenbelasting'\
     'prestatievergoeding rechtspersonenbelasting registratierechten schenkbelasting socialezekerheidsbijdragen solidariteitsbijdrage uitbetalingsinstelling vennootschapsbelasting'\
     'verminderingen vervangingsinkomsten voorafbetalingen voorbelasting voorheffing vrijstellingsregeling waardevermindering werkgeversbijdrage werkingskosten zekerheidsbijdragen'\
     'invaliditeit verzekering'

_tags=tags.split(' ')

real_tag_tensors=[]
for a in _tags:
    input_ids = torch.tensor(tokenizer.encode(a)).unsqueeze(0)  # Batch size 1
    outputs = undisputed_best_model(input_ids)
    real_tag_tensors.append(outputs.logits)
    
    



input_ids = torch.tensor(tokenizer.encode('In de periode tussen 7 en 13 juni zijn dagelijks gemiddeld 60,7 mensen in het ziekenhuis opgenomen omwille van een coronabesmetting, een stijging met 15 procent in vergelijking met de voorgaande periode. Dat blijkt uit recente cijfers van gezondheidsinstituut Sciensano. Ook het aantal besmettingen stijgt.')).unsqueeze(0)  # Batch size 1
outputs = undisputed_best_model(input_ids)
last_hidden_states = outputs[2]  # The last hidden-state is the first element of the output tuple

input_ids2 = torch.tensor(tokenizer.encode('In de periode tussen 7 en 13 juni zijn dagelijks gemiddeld 60,7 mensen in het ziekenhuis opgenomen omwille van een coronabesmetting, een stijging met 15 procent in vergelijking met de voorgaande periode. Dat blijkt uit recente cijfers van gezondheidsinstituut Sciensano. Ook het aantal besmettingen stijgt.')).unsqueeze(0)  # Batch size 1
outputs2 = undisputed_best_model(input_ids2)
last_hidden_states2 = outputs2[2]  # The last hidden-state is the first element of the output tuple



sentence_embedding = torch.mean(outputs[-1], dim=1).squeeze()
print(sentence_embedding)
print(sentence_embedding.size())
sentence_embedding2 = torch.mean(outputs2[-1], dim=1).squeeze()
print(sentence_embedding2)
print(sentence_embedding2.size())



cos = torch.nn.CosineSimilarity(dim=0)

q=cos(sentence_embedding, sentence_embedding2)
q

import spacy
nlp = spacy.load("nl_core_news_lg")
real_tags=nlp(tags)


def tagging(real_tags,summary_tags):
    summary_tag_list={}

    for _a in summary_tags:
        
        for token in real_tags:
            q=round(token.similarity(_a),3)
        
            if q > SIMILARITY:
                
                #add token to dict
                summary_tag_list[token]=q
                #print(_a,token,_a.similarity(token))
    return summary_tag_list

keys=[]
t=0
for a in summary:
    print(t)
    text=a.lower()
    summary_tags=nlp(text)      
    
    summary_tag_list=tagging(real_tags,summary_tags)
    dict1 = summary_tag_list
    sorted_dict = {}
    sorted_keys = sorted(dict1, key=dict1.get,reverse=True)  # [1, 3, 2]

    for w in sorted_keys:
        sorted_dict[w] = dict1[w]

    first5pairs = {k: sorted_dict[k] for k in list(sorted_dict)[:5]}


    keys.append(first5pairs)
    t=t+1

data = pd.DataFrame(
    {'date':DATE,
    'numac':numacs,
    'nltext':nl_list,
        'nllink':numac_links,
        'summary': summary,
     'nltags': keys
    })


keys

import numpy as np
data['nltags']=data['nltags'].astype(str)
data['nltags']=data["nltags"].str.strip('{}')
data['nltags'] = data['nltags'].replace('',np.nan,regex = True)
data.dropna(subset = ["nltags"], inplace=True)
data

#from btax.taxtag.models import Article
for index, row in data.iterrows():
    print(row['date'], row['numac'])

 


# class Article(models.Model):
#     date = models.DateField()
#     numac = models.CharField(max_length=15)
#     link=models.CharField(max_length=150)
#     nl_text=models.TextField()
#     nl_sum=models.TextField()
#     nl_tags=models.TextField()
#     created_at = models.DateTimeField(auto_now_add=True)
#     updated_at = models.DateTimeField(auto_now=True)


list(data['nltags'])

all=list(zip(keys,summary))
